Factor out county wizard start stage.
