+++ client/src/component/county/audit/wizard/CountyAuditWizard.tsx
@@ -2,20 +2,8 @@ import * as React from 'react';
 
 import { Checkbox, EditableText, Radio, RadioGroup } from '@blueprintjs/core';
 
+import WizardStart from './WizardStart';
 
-const WizardStart = ({ nextStage }: any) => (
-    <div>
-        <div>Acme County Audit</div>
-        <div className='pt-card'>
-            <div>Acme County General Election Audit</div>
-            <div>Election date: 11/21/2017</div>
-            <div>County & State Ballot Contests</div>
-            <button className='pt-button' onClick={ nextStage }>
-                Start My Audit
-            </button>
-        </div>
-    </div>
-);
 
 const AuditBoardSignInForm = () => (
     <div>
